Changed abjad.lilypondformat to abjad.format.
Removed abjad.new().
