+++ abjad/score.py
@@ -167,10 +167,10 @@ class Component:
         return []
 
     def _format_component(self, pieces=False):
-        from . import lilypondformat as _lilypondformat
+        from . import format as _format
 
         result = []
-        bundle = _lilypondformat.bundle_format_contributions(self)
+        bundle = _format.bundle_format_contributions(self)
         result.extend(self._format_absolute_before_slot(bundle))
         result.extend(self._format_before_slot(bundle))
         result.extend(self._format_open_brackets_slot(bundle))
@@ -242,11 +242,11 @@ class Component:
         return duration
 
     def _get_format_contributions_for_slot(self, slot_identifier, bundle=None):
-        from . import lilypondformat as _lilypondformat
+        from . import format as _format
 
         result = []
         if bundle is None:
-            bundle = _lilypondformat.bundle_format_contributions(self)
+            bundle = _format.bundle_format_contributions(self)
         slot_names = (
             "before",
             "open_brackets",
